requirements.txt
conductor-python==1.1.4
python-dotenv
boto3==1.34.62
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0